added support for fields that create table aliases that reference the same database table as the parent model.
This is especially useful for foreign key references
that point to the same model a.k.a ForeignKey('self')

The Django ORM will create an alias but this doesn't get
used and so the query fails to perform lookup references
on the actual target field. This commit fixes that by using
whatever aliases were created (which coincidentally aren't
quoted) to perform the lookup.